backend:add restaurant table. add changes for that

+++ backend/src/main/resources/db/migration/V1.2__Add_reservation_request_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `reservation_request` (
   `id` INT NOT NULL AUTO_INCREMENT,
-  `asking_user_id` BIGINT(20) NOT NULL,
-  `reciving_user_id` BIGINT(20) NOT NULL,
+  `asking_user_id` INT NOT NULL,
+  `reciving_user_id` INT NOT NULL,
   `numer_of_persons` INT NOT NULL DEFAULT 0,
   `date_and_time` DATETIME NOT NULL,
   `message` VARCHAR(512) NULL,
